Improve mulit-artist support
The database now sets one album-artist for every track parsed, which is either the value of the albumartist
tag, or the first artist in the 'artist' tag. When selecting an artist-album combo, the album view track
list shows all non-albumartist artists for a track. If the artist selected is not the album-artist, each
track will also list the album-artist.