Change database layout again
Albums table now has an artist column to identify album artists, so that albums can be uniquely identified
by their name and artist, instead of name and date which would cause albums to update incorrectly in the
library when their date was changed. Artists are now identified by the file they were found in, and
the method for identifying and selecting album artists or track artists has been adjusted.