Player controls respond to player updates directly.
Instead of binding volume/mute properties in player to a property in the controls, have the
controls update in response to the notify signal of the player's properties. Also moves code for updating
state of player controls out of the main view so the player controls can connect to the player's signals
directly.