Redo the play queue yet again.
Put it back to a ListView because I figured out how to sync the checkbox with the selection state and
delete selected tracks using a ListView properly. It was quite obvious; don't know how I didn't manage to
figure it out last time I tried this. Now can shift/control select for better selection, but loose the
current track indicator.