Standardize play/add methods
Instead of play/add/replace for both albums and tracks, make a standard set that operates on a list of tracks and use helper methods to connect the callbacks to them. As a consequence, the Play Track and Start Here track options now are consolidated to just Play Track which now adds the entire album to the queue and starts at the specific track.